dev: 메인 홈구장화면 홈구장 데이터, 선수 데이터에 따라 동적 렌더링 반영

+++ lib/model/HomeStadium.dart
@@ -1,10 +1,10 @@
-import 'dart:ffi';
+
 
 class HomeStadium {
   final String? name;
   final Team? team;
   final int? favoriteAthleteCode;
-  final Long? ownerId;
+  final int? ownerId;
 
   HomeStadium({this.name, this.team, this.favoriteAthleteCode, this.ownerId});
 
@@ -12,7 +12,7 @@ class HomeStadium {
     copyWith(name: null, team: null,   ownerId: null);
   }
 
-  HomeStadium copyWith({String? name, Team? team, int? favoriteAthleteCode, Long? ownerId}) =>
+  HomeStadium copyWith({String? name, Team? team, int? favoriteAthleteCode, int? ownerId}) =>
       HomeStadium(
         name: name ?? this.name,
         team: team ?? this.team,
